requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
aiofiles
PyPDF2
python-dotenv
jinja2